Live Dialogue Assistant Web App/requirements.txt
tkinter==8.6
sys==3.8
openai==0.27.0
os==3.8
config==0.5.0.post0
google-cloud-speech==2.13.0
pydub==0.25.1
librosa==0.8.3
soundfile==0.10.3.post1
speechrecognition==3.8.1